+++ src/app/app-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { HomeComponentComponent } from './home-component/home-component.component';
+import { HomeComponentComponent } from './home/home.component';
 
 const routes: Routes = [
   {'path': '', component:HomeComponentComponent},
