--- src/app/home-component/home-component.component.html
@@ -1,128 +1,0 @@
-<head>
-    <meta charset="UTF-8">
-    <meta name="description" content="Welcome to Segala Solutions. We specialize in delivering high-quality technology solutions with honesty, efficiency, and a strong work ethic." />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="icon" type="image/x-icon" href="./images/fav-icon.png">
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css">
-    <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.10.2/dist/umd/popper.min.js"></script>
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@24,400,0,0" />
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Quicksand">
-    
-    <title>Segala Solutions</title>
-  </head>
-  
-  <!-- About Section -->
-  <section id="about" class="about">
-    <div class="row">
-      <div class="col-md-5 banner-statement">
-        <h3>
-          Welcome to Segala Solutions. We specialize in delivering high-quality technology solutions with honesty, efficiency, and a strong work ethic.
-        </h3>
-      </div>
-      <div class="col-md-7">
-        <img class="banner-image" src="./assets/images/banner-image.png" alt="Banner Image">
-      </div>
-    </div>
-    <div class="custom-shape-divider-bottom-1701991771">
-      <svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 1200 120" preserveAspectRatio="none">
-          <path d="M321.39,56.44c58-10.79,114.16-30.13,172-41.86,82.39-16.72,168.19-17.73,250.45-.39C823.78,31,906.67,72,985.66,92.83c70.05,18.48,146.53,26.09,214.34,3V0H0V27.35A600.21,600.21,0,0,0,321.39,56.44Z" class="shape-fill"></path>
-      </svg>
-    </div>
-  </section>
-  
-  <!-- Services Section -->
-  <section id="services" class="services">
-    <h2>Services</h2>
-    <div class="flex-container">
-      <div class="service-tile">
-        <span class="material-symbols-outlined">terminal</span>
-        <p>Software Engineering</p>
-      </div>
-      <div class="service-tile">
-        <span class="material-symbols-outlined">code_blocks</span>
-        <p>Web Development</p>
-      </div>
-      <div class="service-tile">
-        <span class="material-symbols-outlined">cloud</span>
-        <p>Cloud Architecture</p>
-      </div>
-      <div class="service-tile">
-        <span class="material-symbols-outlined">all_inclusive</span>
-        <p>Devops</p>
-      </div>
-      <div class="service-tile">
-        <span class="material-symbols-outlined">manage_accounts</span>
-        <p>Project Management</p>
-      </div>
-    </div>
-  </section>
-  
-  <!-- Technologies Section -->
-  <section id="technologies" class="technologies">
-    <div class="custom-shape-divider-top-1701992275">
-      <svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 1200 120" preserveAspectRatio="none">
-          <path d="M985.66,92.83C906.67,72,823.78,31,743.84,14.19c-82.26-17.34-168.06-16.33-250.45.39-57.84,11.73-114,31.07-172,41.86A600.21,600.21,0,0,1,0,27.35V120H1200V95.8C1132.19,118.92,1055.71,111.31,985.66,92.83Z" class="shape-fill"></path>
-      </svg>
-    </div>
-    <h2>Technologies</h2>
-    <div class="flex-container">
-      <img class="tech-image" src="./assets/images/technology-logos/salesforce-logo.png" alt="Salesforce">
-      <img class="tech-image" src="./assets/images/technology-logos/servicenow-logo.png" alt="Salesforce">
-      <img class="tech-image" src="./assets/images/technology-logos/angular-logo.png" alt="Salesforce">
-      <img class="tech-image" src="./assets/images/technology-logos/react-logo.png" alt="Salesforce">
-      <img class="tech-image" src="./assets/images/technology-logos/vue-logo.png" alt="Salesforce">
-      <img class="tech-image" src="./assets/images/technology-logos/springboot-logo.png" alt="Salesforce">
-      <img class="tech-image" src="./assets/images/technology-logos/aws-logo.png" alt="Salesforce">
-      <img class="tech-image" src="./assets/images/technology-logos/jenkins-logo.png" alt="Salesforce">
-      <img class="tech-image" src="./assets/images/technology-logos/ansible-logo.png" alt="Salesforce">
-    </div>
-    <div class="custom-shape-divider-bottom-1702005791">
-      <svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 1200 120" preserveAspectRatio="none">
-          <path d="M985.66,92.83C906.67,72,823.78,31,743.84,14.19c-82.26-17.34-168.06-16.33-250.45.39-57.84,11.73-114,31.07-172,41.86A600.21,600.21,0,0,1,0,27.35V120H1200V95.8C1132.19,118.92,1055.71,111.31,985.66,92.83Z" class="shape-fill"></path>
-      </svg>
-    </div>
-  </section>
-  
-  <!-- Contact Us Section -->
-  <section id="contact" class="contact-form">
-    <div class="container">
-      <div class="row">
-        <div class="col-md-6 contact-text">
-          <div>
-            <h2>Contact Us</h2>
-            <p class="contact-text-second-line">and we'll be in touch.</p>
-          </div>
-        </div>
-        <div class="col-md-6">
-          <div class="contact-form-wrapper">
-            <form action="https://api.web3forms.com/submit" method="POST" id="form">
-              <input type="hidden" name="access_key" value="d6023f15-d046-4301-bd42-c80bd9b99856" />
-              <input type="hidden" name="redirect" value="https://web3forms.com/success" />
-              <input type="checkbox" name="botcheck" id="" style="display: none;" />
-              <input type="email" name="email" class="form-control" placeholder="Your Email" required>
-              <input type="text" name="subject" class="form-control" placeholder="Subject" required>
-              <textarea name="message" class="form-control" rows="5" placeholder="Message" required></textarea>
-              <button type="submit" class="btn btn-primary">Submit</button>
-            </form>
-          </div>
-        </div>
-      </div>
-    </div>
-    <div class="custom-shape-divider-bottom-1702005911">
-      <svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 1200 120" preserveAspectRatio="none">
-          <path d="M985.66,92.83C906.67,72,823.78,31,743.84,14.19c-82.26-17.34-168.06-16.33-250.45.39-57.84,11.73-114,31.07-172,41.86A600.21,600.21,0,0,1,0,27.35V120H1200V95.8C1132.19,118.92,1055.71,111.31,985.66,92.83Z" class="shape-fill"></path>
-      </svg>
-    </div>
-  </section>
-  
-  <!-- Footer -->
-  <footer class="footer">
-    <div class="container text-center">
-      <img class="footer-logo" src="./assets/images/logo-vertical.png" alt="Segala Solutions Logo">
-      <p>gsegala@segalasolutions.com | 410-294-0545</p>
-    </div>
-  </footer>
-  
-  <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"></script>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"></script>
-  <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"></script>
